+++ source/background/library/messaging.js
@@ -1,9 +1,5 @@
 import { extractDomain } from "../../shared/library/domain.js";
-import {
-    consumeVaultFacade,
-    createEntryFacade,
-    createVaultFacade
-} from "../../shared/library/buttercup.js";
+import { consumeVaultFacade, createEntryFacade, createVaultFacade } from "../../shared/library/buttercup.js";
 import { dispatch, getState } from "../redux/index.js";
 import log from "../../shared/library/log.js";
 import {
@@ -14,6 +10,7 @@ import {
     generateEntryPath,
     getArchive,
     getEntry,
+    getFacades,
     getNameForSource,
     getSourceIDForVaultID,
     getSourcesInfo,
@@ -375,6 +372,7 @@ function handleMessage(request, sender, sendResponse) {
                     value: request.value,
                 })
             );
+            return false;
         }
         case "set-generated-password": {
             const { password } = request;
@@ -416,17 +414,22 @@ function handleMessage(request, sender, sendResponse) {
 }
 
 async function processSearchResults([entryResults, sources]) {
+    const sourceIDs = {};
+    const sourceNames = {};
+    const vaultFacades = await getFacades();
     const results = await Promise.all(
         entryResults.map(async entryResult => {
-            const sourceID = await getSourceIDForVaultID(entryResult.vaultID);
-            const sourceName = await getNameForSource(sourceID);
-            const entry = await getEntry(sourceID, entryResult.id);
-            const facade = createEntryFacade(entry);
+            const sourceID = (sourceIDs[entryResult.vaultID] =
+                sourceIDs[entryResult.vaultID] || (await getSourceIDForVaultID(entryResult.vaultID)));
+            const sourceName = (sourceNames[entryResult.vaultID] =
+                sourceNames[entryResult.vaultID] || (await getNameForSource(sourceID)));
+            const vaultFacade = vaultFacades.find(facade => facade.sourceID === sourceID);
+            const entryFacade = vaultFacade.entries.find(e => e.id === entryResult.id);
             return {
                 title: entryResult.properties.title,
                 id: entryResult.id,
-                entryPath: generateEntryPath(entry),
-                facade,
+                entryPath: generateEntryPath(vaultFacade, entryResult.id),
+                facade: entryFacade,
                 sourceID,
                 sourceName,
                 url: entryResult.urls[0] || null,