+++ source/dialog/redux/index.js
@@ -2,8 +2,8 @@ import { applyMiddleware, createStore, compose } from "redux";
 import thunk from "redux-thunk";
 import { routerMiddleware } from "connected-react-router";
 import { createSyncMiddleware, syncStore } from "redux-browser-extension-sync";
-import createDialogReducer from "../reducers/index.js";
-import history from "./history.js";
+import createDialogReducer from "../reducers";
+import history from "./history";
 
 const reduxRouterMiddleware = routerMiddleware(history);
 const syncMiddleware = createSyncMiddleware();