+++ source/setup/containers/VaultEditor.js
@@ -14,6 +14,7 @@ import {
 import { notifyError, notifySuccess } from "../library/notify.js";
 import { setBusy, unsetBusy } from "../../shared/actions/app.js";
 import { arrayBufferToBase64, base64ToArrayBuffer } from "../../shared/library/buffer.js";
+import { getConfigKey } from "../../shared/selectors/app.js";
 
 async function convertAttachmentFiles(files) {
     const output = [];
@@ -31,6 +32,7 @@ async function convertAttachmentFiles(files) {
 export default withTranslation()(
     connect(
         (state, ownProps) => ({
+            dynamicIconsSetting: getConfigKey(state, "dynamicIcons"),
             vault: getVaultFacade(state),
         }),
         {