+++ source/shared/library/buttercup.js
@@ -1,3 +1,3 @@
-const { init } = (module.exports = require(_BUTTERCUP_WEB_IMPORT_));
+const { init } = (module.exports = require("buttercup/web"));
 
 init();