+++ source/shared/library/config.js
@@ -1,6 +1,7 @@
 export const INITIAL_CONFIG = {
     language: "en",
     darkMode: false,
+    dynamicIcons: "enabled",
     showSaveDialog: "always", // always/unlocked/never
     autoUnlockVaults: true,
     autoLockVaults: "off",